cleaned up redux-related code structure
