+++ src/client/src/configureStore.ts
@@ -7,7 +7,7 @@ import rootReducer from './redux/reducers/combineReducers'
 import { blotterEpic } from './ui/blotter/'
 import { referenceServiceEpic } from './redux/actions/referenceDataActions'
 import { pricingServiceEpic } from './redux/actions/pricingActions'
-import { analyticsServiceEpic } from './ui/analytics/analyticsOperations'
+import { analyticsServiceEpic } from './ui/analytics'
 import { compositeStatusServiceEpic } from './redux/reducers/compositeStatusServiceOperations'
 import { connectionStatusEpicsCreator } from './connectionStatusOperations'
 import { popoutEpic } from './popoutEpic'
