+++ src/client/src/ui/analytics/index.ts
@@ -1,5 +1,7 @@
-export { default } from './Analytics'
+export { default as Analytics } from './Analytics'
 export { default as AnalyticsContainer } from './AnalyticsContainer'
 export { default as PNLChart } from './PNLChart'
 export { default as PositionsBubbleChart } from './positions-chart/PositionsBubbleChart'
 export { default as AnalyticsBarChart } from './chart/AnalyticsBarChart'
+export { default as analyticsReducer } from './reducer'
+export { default as analyticsServiceEpic} from './epics'
